三平面の描画方法を変更

+++ src/SolidEditForm.h
@@ -62,7 +62,7 @@ public:
     void initializeGL();              //  OpenGL初期化
     void resizeGL(int, int);          //  ウィジットリサイズ時のハンドラ
     void paintGL();                   //  描画処理
-    void paintObject(CObject* obj,QVector<float> color,int tick); //  選択したオブジェクトを描画する
+    void paintObject(CObject* obj, QVector<double> color, int tick); //  選択したオブジェクトを描画する
 
 public slots:
